Revert "Merge remote-tracking branch 'origin/release-35.x' into develop"
This reverts commit 99db33dcc78df638159a2a3d9d1b0a98eda0b772.

+++ features/notifications/ticket-strategy/src/main/java/org/opennms/netmgt/notifd/TicketNotificationStrategy.java
@@ -144,7 +144,7 @@ public class TicketNotificationStrategy implements NotificationStrategy {
         }
         
         // We know the event is an alarm, pull the alarm and current ticket details from the database
-        AlarmState alarmState = getAlarmStateFromEvent(Long.parseLong(eventID));
+        AlarmState alarmState = getAlarmStateFromEvent(Integer.parseInt(eventID));
         if( alarmState.getAlarmID() == 0 ) {
 		LOG.error("There is no alarm-id associated with the event-id='{}'. Will not create ticket.", eventID);
         	return 1;
@@ -165,7 +165,7 @@ public class TicketNotificationStrategy implements NotificationStrategy {
      *
      * @return 0 if alarmid is null
      */
-	protected AlarmState getAlarmStateFromEvent(long eventID) {
+	protected AlarmState getAlarmStateFromEvent(int eventID) {
 		AlarmStateRowCallbackHandler callbackHandler = new AlarmStateRowCallbackHandler();
 
         JdbcTemplate template = new JdbcTemplate(DataSourceFactory.getInstance());
